ZA: Move the elections related classes in views.py to their own file
